% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_ratio.R
\name{plot_tree_ratio}
\alias{plot_tree_ratio}
\title{Plot additive effects as result of change decomposition for ratio indicators}
\usage{
plot_tree_ratio(tree, facet_labels = c("Additive effects", "Ratio",
  "Numerator", "Denominator"), eff_labels = c("Total effect",
  "Weight effect", "Ratio effect", "Residual effect"),
  root_num_denom = FALSE, col_pal = NULL, ...)
}
\arguments{
\item{tree}{a tree object with all decomposed effect precomputed}

\item{facet_labels}{custom user-defined labels for chart columns}

\item{eff_labels}{custom user-defined labels for guide}

\item{root_num_denom}{logical on/off switch for ploting root node's numerator and denominator}

\item{col_pal}{custom user-defined collor pallette}

\item{...}{args passed to data.tree::ToDataFrameTree}
}
\description{
Plot additive effects as result of change decomposition for ratio indicators
}
\details{
\code{plot_tree_ratio()} requires \code{ggplot2} package
}
\examples{
plot_tree_ratio(tree)
}
